moves mock attacker to a util file
